1.0.170331.1 対応＋PaaS 対応

+++ Deploy/Set-VHDBoot.ps1
@@ -27,15 +27,14 @@ Param
     [Parameter(Mandatory=$True,Position=1,ParameterSetName='Manual')]
     [string]$VHDXFullPath,
     # cloudbuilder.vhdx folder
-    [Parameter(Mandatory=$True,Position=2,ParameterSetName='Manual')]
-    [string]$MasterVHDXFullPath ,
+    [Parameter(Mandatory=$false,Position=2,ParameterSetName='Manual')]
+    [string]$MasterVHDXFullPath = "none",
     # cloudbuilder.vhdx folder
     [Parameter(Mandatory=$false,Position=3,ParameterSetName='Manual')]
     [string]$VHDXPath = 'C:\mas'
 )
 $Error.Clear();
 $ErrorActionPreference = 'Stop';
-
 
 if ($PsCmdlet.ParameterSetName -eq "File")
 {
@@ -54,22 +53,42 @@ if ($PsCmdlet.ParameterSetName -eq "File")
 }
 
 # Create folder
-Write-Host "Create folder : $VHDXPath" -ForegroundColor Green;
-New-Item $VHDXPath -Type directory -Force | Out-Null;
+if (Test-Path -Path $VHDXPath)
+{
+    Write-Host "skip Create folder : $VHDXPath" -ForegroundColor Green;
+}
+else
+{
+    Write-Host "Create folder : $VHDXPath" -ForegroundColor Green;
+    New-Item $VHDXPath -Type directory -Force | Out-Null;
+}
 
 # Download files
 Write-Host "Download files" -ForegroundColor Green;
 $Uri = 'https://raw.githubusercontent.com/Azure/AzureStack-Tools/master/Deployment/';
 'BootMenuNoKVM.ps1', 'PrepareBootFromVHD.ps1', 'unattend_NoKVM.xml' | ForEach-Object { Invoke-WebRequest ($uri + $_) -OutFile ($VHDXPath + '\' + $_) } ;
 
-# copy CloudBuilder
-Write-Host "Copy vhdx" -ForegroundColor Green;
-Write-Verbose " Source Path : $MasterVHDXFullPath";
-Write-Verbose " Destination Path : $VHDXPath";
-Copy-Item -Path $MasterVHDXFullPath -Destination $VHDXPath -Force;
+if ($MasterVHDXFullPath -ne "none")
+{
+    if (Test-Path $MasterVHDXFullPath)
+    {
+        # copy CloudBuilder
+        Write-Host "Copy vhdx" -ForegroundColor Green;
+        Write-Verbose " Source Path : $MasterVHDXFullPath";
+        Write-Verbose " Destination Path : $VHDXPath";
+        Copy-Item -Path $MasterVHDXFullPath -Destination $VHDXPath -Force;        
+    }
+    else
+    {
+        Write-Error " file not found ";
+        return ;        
+    }
+}
+else
+{
+    Write-Host "skip Copy vhdx" -ForegroundColor Green;
+}
 
-# change location
-Write-Host "Change location : $VHDXPath" -ForegroundColor Green;
-Set-Location $VHDXPath;
+# Set vhd boot
 Write-Host "Set vhd boot" -ForegroundColor Green;
-.\PrepareBootFromVHD.ps1 -CloudBuilderDiskPath $VHDXFullPath -ApplyUnattend -AdminPassword $HostPassword -Verbose
+& "$VHDXPath\PrepareBootFromVHD.ps1" -CloudBuilderDiskPath $VHDXFullPath -ApplyUnattend -AdminPassword $HostPassword -Verbose
